Rearrange and rename
