+++ wsidicom/wsidicom.py
@@ -12,14 +12,10 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-import math
 import os
 import warnings
-from abc import ABCMeta, abstractmethod
-from collections import defaultdict
 from pathlib import Path
-from typing import (Callable, DefaultDict, Dict, List, Optional, OrderedDict,
-                    Sequence, Tuple, Union, cast)
+from typing import Callable, List, Optional, Sequence, Tuple, Union
 
 from PIL import Image
 from pydicom.dataset import FileMetaDataset
@@ -27,1425 +23,17 @@ from pydicom.filereader import read_file_meta_info
 from pydicom.misc import is_dicom
 from pydicom.uid import UID, generate_uid
 
+from wsidicom.dataset import WsiDicomDataset
 from wsidicom.errors import (WsiDicomMatchError, WsiDicomNotFoundError,
                              WsiDicomOutOfBoundsError)
+from wsidicom.file import WsiDicomFile
 from wsidicom.geometry import Point, PointMm, Region, RegionMm, Size, SizeMm
 from wsidicom.graphical_annotations import AnnotationInstance
-from wsidicom.instance import (ImageData, ImageOrgin, WsiDataset, WsiDicomFile,
-                               WsiDicomFileWriter, WsiDicomImageData,
-                               WsiInstance)
+from wsidicom.instance import Instance, Level
 from wsidicom.optical import OpticalManager
-from wsidicom.stringprinting import (dict_pretty_str, list_pretty_str,
-                                     str_indent)
+from wsidicom.series import Labels, Levels, Overviews, Series
+from wsidicom.stringprinting import list_pretty_str
 from wsidicom.uid import ANN_SOP_CLASS_UID, WSI_SOP_CLASS_UID, SlideUids
-
-
-class WsiDicomGroup:
-    """Represents a group of instances having the same size, but possibly
-    different z coordinate and/or optical path."""
-    def __init__(
-        self,
-        instances: Sequence[WsiInstance]
-    ):
-        """Create a group of WsiInstances. Instances should match in the common
-        uids, wsi type, and tile size.
-
-        Parameters
-        ----------
-        instances: Sequence[WsiInstance]
-            Instances to build the group.
-        """
-        self._instances = {  # key is identifier (Uid)
-            instance.identifier: instance for instance in instances
-        }
-        self._validate_group()
-
-        base_instance = instances[0]
-        self._wsi_type = base_instance.wsi_type
-        self._uids = base_instance.uids
-
-        self._size = base_instance.size
-        self._pixel_spacing = base_instance.pixel_spacing
-        self._default_instance_uid: UID = base_instance.identifier
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}({self.instances.values()})"
-
-    def __str__(self) -> str:
-        return self.pretty_str()
-
-    def pretty_str(
-        self,
-        indent: int = 0,
-        depth: Optional[int] = None
-    ) -> str:
-        string = (
-            f'Image: size: {self.size} px, mpp: {self.mpp} um/px'
-        )
-        if depth is not None:
-            depth -= 1
-            if(depth < 0):
-                return string
-        string += (
-            ' Instances: ' + dict_pretty_str(self.instances, indent, depth)
-        )
-        return string
-
-    def __getitem__(self, index: UID) -> WsiInstance:
-        return self.instances[index]
-
-    @property
-    def uids(self) -> SlideUids:
-        """Return uids"""
-        return self._uids
-
-    @property
-    def wsi_type(self) -> str:
-        """Return wsi type"""
-        return self._wsi_type
-
-    @property
-    def size(self) -> Size:
-        """Return image size in pixels"""
-        return self._size
-
-    @property
-    def mpp(self) -> SizeMm:
-        """Return pixel spacing in um/pixel"""
-        return self.pixel_spacing*1000.0
-
-    @property
-    def pixel_spacing(self) -> SizeMm:
-        """Return pixel spacing in mm/pixel"""
-        return self._pixel_spacing
-
-    @property
-    def instances(self) -> Dict[UID, WsiInstance]:
-        """Return contained instances"""
-        return self._instances
-
-    @property
-    def default_instance(self) -> WsiInstance:
-        """Return default instance"""
-        return self.instances[self._default_instance_uid]
-
-    @property
-    def files(self) -> List[Path]:
-        """Return contained files"""
-        instance_files = [
-            instance.image_data.files for instance in self.instances.values()
-        ]
-        return [file for sublist in instance_files for file in sublist]
-
-    @property
-    def datasets(self) -> List[WsiDataset]:
-        """Return contained datasets."""
-        instance_datasets = [
-            instance.datasets for instance in self.instances.values()
-        ]
-        return [
-            dataset for sublist in instance_datasets for dataset in sublist
-        ]
-
-    @property
-    def optical_paths(self) -> List[str]:
-        return list({
-            path
-            for instance in self.instances.values()
-            for path in instance.optical_paths
-        })
-
-    @property
-    def focal_planes(self) -> List[float]:
-        return list({
-            focal_plane
-            for innstance in self.instances.values()
-            for focal_plane in innstance.focal_planes
-        })
-
-    @property
-    def image_origin(self) -> ImageOrgin:
-        return self.default_instance.image_origin
-
-    @classmethod
-    def open(
-        cls,
-        instances: Sequence[WsiInstance],
-    ) -> List['WsiDicomGroup']:
-        """Return list of groups created from wsi instances.
-
-        Parameters
-        ----------
-        files: Sequence[WsiInstance]
-            Instances to create groups from.
-
-        Returns
-        ----------
-        List[WsiDicomGroup]
-            List of created groups.
-
-        """
-        groups: List['WsiDicomGroup'] = []
-
-        grouped_instances = cls._group_instances(instances)
-
-        for group in grouped_instances.values():
-            groups.append(cls(group))
-
-        return groups
-
-    def matches(self, other_group: 'WsiDicomGroup') -> bool:
-        """Check if group matches other group. If strict common Uids should
-        match. Wsi type should always match.
-
-        Parameters
-        ----------
-        other_group: WsiDicomGroup
-            Other group to match against.
-
-        Returns
-        ----------
-        bool
-            True if other group matches.
-        """
-        return (
-            self.uids.matches(other_group.uids) and
-            other_group.wsi_type == self.wsi_type
-        )
-
-    def valid_pixels(self, region: Region) -> bool:
-        """Check if pixel region is withing the size of the group image size.
-
-        Parameters
-        ----------
-        region: Region
-            Pixel region to check
-
-        Returns
-        ----------
-        bool
-            True if pixel position and size is within image
-        """
-        # Return true if inside pixel plane.
-        image_region = Region(Point(0, 0), self.size)
-        return region.is_inside(image_region)
-
-    def get_instance(
-        self,
-        z: Optional[float] = None,
-        path: Optional[str] = None
-    ) -> WsiInstance:
-        """Search for instance fullfilling the parameters.
-        The behavior when z and/or path is none could be made more
-        clear.
-
-        Parameters
-        ----------
-        z: Optional[float] = None
-            Z coordinate of the searched instance
-        path: Optional[str] = None
-            Optical path of the searched instance
-
-        Returns
-        ----------
-        WsiInstance
-            The instance containing selected path and z coordinate
-        """
-        if z is None and path is None:
-            instance = self.default_instance
-            z = instance.default_z
-            path = instance.default_path
-
-            return self.default_instance
-
-        # Sort instances by number of focal planes (prefer simplest instance)
-        sorted_instances = sorted(
-            list(self._instances.values()),
-            key=lambda i: len(i.focal_planes)
-        )
-        try:
-            if z is None:
-                # Select the instance with selected optical path
-                instance = next(i for i in sorted_instances if
-                                path in i.optical_paths)
-            elif path is None:
-                # Select the instance with selected z
-                instance = next(i for i in sorted_instances
-                                if z in i.focal_planes)
-            else:
-                # Select by both path and z
-                instance = next(i for i in sorted_instances
-                                if (z in i.focal_planes and
-                                    path in i.optical_paths))
-        except StopIteration:
-            raise WsiDicomNotFoundError(
-                f"Instance for path: {path}, z: {z}", "group"
-            )
-        if z is None:
-            z = instance.default_z
-        if path is None:
-            path = instance.default_path
-        return instance
-
-    def get_default_full(self) -> Image.Image:
-        """Read full image using default z coordinate and path.
-
-        Returns
-        ----------
-        Image.Image
-            Full image of the group.
-        """
-        instance = self.default_instance
-        z = instance.default_z
-        path = instance.default_path
-        region = Region(position=Point(x=0, y=0), size=self.size)
-        image = self.get_region(region, z, path)
-        return image
-
-    def get_region(
-        self,
-        region: Region,
-        z: Optional[float] = None,
-        path: Optional[str] = None,
-    ) -> Image.Image:
-        """Read region defined by pixels.
-
-        Parameters
-        ----------
-        location: int, int
-            Upper left corner of region in pixels
-        size: int
-            Size of region in pixels
-        z: Optional[float] = None
-            Z coordinate, optional
-        path: Optional[str] = None
-            optical path, optional
-
-        Returns
-        ----------
-        Image.Image
-            Region as image
-        """
-
-        instance = self.get_instance(z, path)
-        if z is None:
-            z = instance.default_z
-        if path is None:
-            path = instance.default_path
-        image = instance.image_data.stitch_tiles(region, path, z)
-        return image
-
-    def get_region_mm(
-        self,
-        region: RegionMm,
-        z: Optional[float] = None,
-        path: Optional[str] = None,
-        slide_origin: bool = False
-    ) -> Image.Image:
-        """Read region defined by mm.
-
-        Parameters
-        ----------
-        region: RegionMm
-            Region defining upper left corner and size in mm.
-        z: Optional[float] = None
-            Z coordinate, optional.
-        path: Optional[str] = None
-            optical path, optional.
-        slide_origin: bool = False.
-            If to use the slide origin instead of image origin.
-
-        Returns
-        ----------
-        Image.Image
-            Region as image
-        """
-        if slide_origin:
-            region = self.image_origin.transform_region(region)
-        pixel_region = self.mm_to_pixel(region)
-        image = self.get_region(pixel_region, z, path)
-        if slide_origin:
-            image = image.rotate(
-                self.image_origin.rotation,
-                resample=Image.Resampling.BILINEAR,
-                expand=True
-            )
-        return image
-
-    def get_tile(
-        self,
-        tile: Point,
-        z: Optional[float] = None,
-        path: Optional[str] = None
-    ) -> Image.Image:
-        """Return tile at tile coordinate x, y as image.
-
-        Parameters
-        ----------
-        tile: Point
-            Tile x, y coordinate
-        z: Optional[float] = None
-            Z coordinate
-        path: Optional[str] = None
-            Optical path
-
-        Returns
-        ----------
-        Image.Image
-            The tile as image
-        """
-
-        instance = self.get_instance(z, path)
-        if z is None:
-            z = instance.default_z
-        if path is None:
-            path = instance.default_path
-        return instance.image_data.get_tile(tile, z, path)
-
-    def get_encoded_tile(
-        self,
-        tile: Point,
-        z: Optional[float] = None,
-        path: Optional[str] = None
-    ) -> bytes:
-        """Return tile at tile coordinate x, y as bytes.
-
-        Parameters
-        ----------
-        tile: Point
-            Tile x, y coordinate
-        z: Optional[float] = None
-            Z coordinate
-        path: Optional[str] = None
-            Optical path
-
-        Returns
-        ----------
-        bytes
-            The tile as bytes
-        """
-        instance = self.get_instance(z, path)
-        if z is None:
-            z = instance.default_z
-        if path is None:
-            path = instance.default_path
-        return instance.image_data.get_encoded_tile(tile, z, path)
-
-    def mm_to_pixel(self, region: RegionMm) -> Region:
-        """Convert region in mm to pixel region.
-
-        Parameters
-        ----------
-        region: RegionMm
-            Region in mm
-
-        Returns
-        ----------
-        Region
-            Region in pixels
-        """
-        pixel_region = Region(
-            position=region.position // self.pixel_spacing,
-            size=region.size // self.pixel_spacing
-        )
-        if not self.valid_pixels(pixel_region):
-            raise WsiDicomOutOfBoundsError(
-                f"Region {region}", f"level size {self.size}"
-            )
-        return pixel_region
-
-    def close(self) -> None:
-        """Close all instances on the group."""
-        for instance in self._instances.values():
-            instance.close()
-
-    def _validate_group(self):
-        """Check that no file or instance in group is duplicate, and if strict
-        instances in group matches. Raises WsiDicomMatchError otherwise.
-        """
-        instances = list(self.instances.values())
-        base_instance = instances[0]
-        for instance in instances[1:]:
-            if not base_instance.matches(instance):
-                raise WsiDicomMatchError(str(instance), str(self))
-
-        WsiDataset.check_duplicate_dataset(self.datasets, self)
-        WsiInstance.check_duplicate_instance(instances, self)
-
-    @classmethod
-    def _group_instances(
-        cls,
-        instances: Sequence[WsiInstance]
-    ) -> OrderedDict[Size, List[WsiInstance]]:
-        """Return instances grouped and sorted by image size.
-
-        Parameters
-        ----------
-        instances: Sequence[WsiInstance]
-            Instances to group by image size.
-
-        Returns
-        ----------
-        OrderedDict[Size, List[WsiInstance]]:
-            Instances grouped by size, with size as key.
-
-        """
-        grouped_instances: Dict[Size, List[WsiInstance]] = {}
-        for instance in instances:
-            try:
-                grouped_instances[instance.size].append(instance)
-            except KeyError:
-                grouped_instances[instance.size] = [instance]
-        return OrderedDict(sorted(
-            grouped_instances.items(),
-            key=lambda item: item[0].width,
-            reverse=True)
-        )
-
-    def _group_instances_to_file(
-        self,
-    ) -> List[List[WsiInstance]]:
-        """Group instances by properties that can't differ in a DICOM-file,
-        i.e. the instances are grouped by output file.
-
-        Returns
-        ----------
-        List[List[WsiInstance]]
-            Instances grouped by common properties.
-        """
-        groups: DefaultDict[
-            Tuple[str, UID, bool, Optional[int], Optional[float], str],
-            List[WsiInstance]
-        ] = defaultdict(list)
-
-        for instance in self.instances.values():
-            groups[
-                instance.image_data.photometric_interpretation,
-                instance.image_data.transfer_syntax,
-                instance.ext_depth_of_field,
-                instance.ext_depth_of_field_planes,
-                instance.ext_depth_of_field_plane_distance,
-                instance.focus_method
-            ].append(
-                instance
-            )
-        return list(groups.values())
-
-    @staticmethod
-    def _list_image_data(
-        instances: Sequence[WsiInstance]
-    ) -> Dict[Tuple[str, float], ImageData]:
-        """Sort ImageData in instances by optical path and focal
-        plane.
-
-        Parameters
-        ----------
-        instances: Sequence[WsiInstance]
-            List of instances with optical paths and focal planes to list and
-            sort.
-
-        Returns
-        ----------
-        Dict[Tuple[str, float], ImageData]:
-            ImageData sorted by optical path and focal plane.
-        """
-        output: Dict[Tuple[str, float], ImageData] = {}
-        for instance in instances:
-            for optical_path in instance.optical_paths:
-                for z in sorted(instance.focal_planes):
-                    if (optical_path, z) not in output:
-                        output[optical_path, z] = instance.image_data
-        return output
-
-    def save(
-        self,
-        output_path: str,
-        uid_generator: Callable[..., UID],
-        workers: int,
-        chunk_size: int,
-        offset_table: Optional[str]
-    ) -> List[Path]:
-        """Save a WsiDicomGroup to files in output_path. Instances are grouped
-        by properties that cant differ in the same file:
-            - photometric interpretation
-            - transfer syntax
-            - extended depth of field (and planes and distance)
-            - focus method
-        Other properties are assumed to be equal or to be updated.
-
-        Parameters
-        ----------
-        output_path: str
-            Folder path to save files to.
-        uid_generator: Callable[..., UID]
-            Uid generator to use.
-        workers: int
-            Maximum number of thread workers to use.
-        chunk_size: int
-            Chunk size (number of tiles) to process at a time. Actual chunk
-            size also depends on minimun_chunk_size from image_data.
-        offset_table: Optional[str]
-            Offset table to use, 'bot' basic offset table, 'eot' extended
-            offset table, None - no offset table.
-
-        Returns
-        ----------
-        List[str]
-            List of paths of created files.
-        """
-        filepaths: List[Path] = []
-        for instances in self._group_instances_to_file():
-            uid = uid_generator()
-            filepath = Path(os.path.join(output_path, uid + '.dcm'))
-            transfer_syntax = instances[0].image_data.transfer_syntax
-            image_data_list = self._list_image_data(instances)
-            dataset = instances[0].dataset.as_tiled_full(image_data_list)
-            with WsiDicomFileWriter(filepath) as wsi_file:
-                wsi_file.write(
-                    uid,
-                    transfer_syntax,
-                    dataset,
-                    image_data_list,
-                    workers,
-                    chunk_size,
-                    offset_table
-                )
-            filepaths.append(filepath)
-        return filepaths
-
-
-class WsiDicomLevel(WsiDicomGroup):
-    """Represents a level in the pyramid and contains one or more instances
-    having the same pyramid level index, pixel spacing, and size but possibly
-    different focal planes and/or optical paths.
-    """
-    def __init__(
-        self,
-        instances: Sequence[WsiInstance],
-        base_pixel_spacing: SizeMm
-    ):
-        """Create a level from list of WsiInstances. Asign the pyramid level
-        index from pixel spacing of base level.
-
-        Parameters
-        ----------
-        instances: Sequence[WsiInstance]
-            Instances to build the level.
-        base_pixel_spacing: SizeMm
-            Pixel spacing of base level.
-        """
-        super().__init__(instances)
-        self._base_pixel_spacing = base_pixel_spacing
-        self._level = self._assign_level(self._base_pixel_spacing)
-
-    def __repr__(self) -> str:
-        return (
-            f"{type(self).__name__}({self.instances}, "
-            f"{self._base_pixel_spacing})"
-        )
-
-    def __str__(self) -> str:
-        return self.pretty_str()
-
-    def pretty_str(
-        self,
-        indent: int = 0,
-        depth: Optional[int] = None
-    ) -> str:
-        string = (
-            f'Level: {self.level}, size: {self.size} px, mpp: {self.mpp} um/px'
-        )
-        if depth is not None:
-            depth -= 1
-            if(depth < 0):
-                return string
-        string += (
-            ' Instances: ' + dict_pretty_str(self.instances, indent, depth)
-        )
-        return string
-
-    @property
-    def pyramid(self) -> str:
-        """Return string representation of the level"""
-        return (
-            f'Level [{self.level}]'
-            f' tiles: {self.default_instance.tiled_size},'
-            f' size: {self.size}, mpp: {self.mpp} um/px'
-        )
-
-    @property
-    def tile_size(self) -> Size:
-        return self.default_instance.tile_size
-
-    @property
-    def level(self) -> int:
-        """Return pyramid level"""
-        return self._level
-
-    @classmethod
-    def open(
-        cls,
-        instances: Sequence[WsiInstance],
-    ) -> List['WsiDicomLevel']:
-        """Return list of levels created wsi files.
-
-        Parameters
-        ----------
-        files: Sequence[WsiInstance]
-            Instances to create levels from.
-
-        Returns
-        ----------
-        List[WsiDicomLevel]
-            List of created levels.
-
-        """
-        levels: List['WsiDicomLevel'] = []
-        instances_grouped_by_level = cls._group_instances(instances)
-        base_group = list(instances_grouped_by_level.values())[0]
-        base_pixel_spacing = base_group[0].pixel_spacing
-        for level in instances_grouped_by_level.values():
-            levels.append(cls(level, base_pixel_spacing))
-        return levels
-
-    def matches(self, other_level: 'WsiDicomGroup') -> bool:
-        """Check if level matches other level. If strict common Uids should
-        match. Wsi type and tile size should always match.
-
-        Parameters
-        ----------
-        other_level: WsiDicomGroup
-            Other level to match against.
-
-        Returns
-        ----------
-        bool
-            True if other level matches.
-        """
-        other_level = cast(WsiDicomLevel, other_level)
-        return (
-            self.uids.matches(other_level.uids) and
-            other_level.wsi_type == self.wsi_type and
-            other_level.tile_size == self.tile_size
-        )
-
-    def get_highest_level(self) -> int:
-        """Return highest deep zoom scale that can be produced
-        from the image in the level.
-
-        Returns
-        ----------
-        int
-            Relative level where the pixel size becomes 1x1
-        """
-        return math.ceil(math.log2(max(self.size.width, self.size.height)))
-
-    def get_scaled_tile(
-        self,
-        tile: Point,
-        level: int,
-        z: Optional[float] = None,
-        path: Optional[str] = None
-    ) -> Image.Image:
-        """Return tile in another level by scaling a region.
-        If the tile is an edge tile, the resulting tile is croped
-        to remove part outside of the image (as defiend by level size).
-
-        Parameters
-        ----------
-        tile: Point
-            Non scaled tile coordinate
-        level: int
-            Level to scale from
-        z: Optional[float] = None
-            Z coordinate
-        path: Optional[str] = None
-            Optical path
-
-        Returns
-        ----------
-        Image.Image
-            A tile image
-        """
-        scale = self.calculate_scale(level)
-        instance = self.get_instance(z, path)
-        scaled_region = Region.from_tile(tile, instance.tile_size) * scale
-        cropped_region = scaled_region.crop(instance.image_data.image_size)
-        if not self.valid_pixels(cropped_region):
-            raise WsiDicomOutOfBoundsError(
-                f"Region {cropped_region}", f"level size {self.size}"
-            )
-        image = self.get_region(cropped_region, z, path)
-        tile_size = cropped_region.size.ceil_div(scale)
-        image = image.resize(
-            tile_size.to_tuple(),
-            resample=Image.Resampling.BILINEAR
-        )
-        return image
-
-    def get_scaled_encoded_tile(
-        self,
-        tile: Point,
-        scale: int,
-        z: Optional[float] = None,
-        path: Optional[str] = None
-    ) -> bytes:
-        """Return encoded tile in another level by scaling a region.
-
-        Parameters
-        ----------
-        tile: Point
-            Non scaled tile coordinate
-        level: int
-           Level to scale from
-        z: Optional[float] = None
-            Z coordinate
-        path: Optional[str] = None
-            Optical path
-
-        Returns
-        ----------
-        bytes
-            A transfer syntax encoded tile
-        """
-        image = self.get_scaled_tile(tile, scale, z, path)
-        instance = self.get_instance(z, path)
-        return instance.image_data.encode(image)
-
-    def calculate_scale(self, level_to: int) -> int:
-        """Return scaling factor to given level.
-
-        Parameters
-        ----------
-        level_to -- index of level to scale to
-
-        Returns
-        ----------
-        int
-            Scaling factor between this level and given level
-        """
-        return int(2 ** (level_to - self.level))
-
-    def _assign_level(self, base_pixel_spacing: SizeMm) -> int:
-        """Return (2^level scale factor) based on pixel spacing.
-        Will round to closest integer. Raises NotImplementedError if level is
-        to far from integer.
-
-        Parameters
-        ----------
-        base_pixel_spacing: SizeMm
-            The pixel spacing of the base lavel
-
-        Returns
-        ----------
-        int
-            The pyramid order of the level
-        """
-        float_level = math.log2(
-            self.pixel_spacing.width/base_pixel_spacing.width
-        )
-        level = int(round(float_level))
-        TOLERANCE = 1e-2
-        if not math.isclose(float_level, level, rel_tol=TOLERANCE):
-            raise NotImplementedError("Levels needs to be integer.")
-        return level
-
-    def create_child(
-        self,
-        scale: int,
-        output_path: Path,
-        uid_generator: Callable[..., UID],
-        workers: int,
-        chunk_size: int,
-        offset_table: Optional[str]
-    ) -> 'WsiDicomLevel':
-        """Creates a new WsiDicomLevel from this level by scaling the image
-        data.
-
-        Parameters
-        ----------
-        scale: int
-            Scale factor.
-        output_path: Path
-            The path to write child to.
-        uid_generator: Callable[..., UID]
-            Uid generator to use.
-        workers: int
-            Maximum number of thread workers to use.
-        chunk_size: int
-            Chunk size (number of tiles) to process at a time. Actual chunk
-            size also depends on minimun_chunk_size from image_data.
-        offset_table: Optional[str]
-            Offset table to use, 'bot' basic offset table, 'eot' extended
-            offset table, None - no offset table.
-
-        Returns
-        ----------
-        'WsiDicomLevel'
-            Created scaled level.
-        """
-        filepaths: List[Path] = []
-        if not isinstance(scale, int) or scale < 2:
-            raise ValueError(
-                "Scale must be integer and larger than 2"
-            )
-        if not isinstance(
-            self.default_instance.image_data,
-            WsiDicomImageData
-        ):
-            raise NotImplementedError(
-                "Can only construct pyramid from DICOM WSI files"
-            )
-
-        for instances in self._group_instances_to_file():
-            uid = uid_generator()
-            filepath = Path(os.path.join(output_path, uid + '.dcm'))
-            transfer_syntax = instances[0].image_data.transfer_syntax
-            image_data_list = self._list_image_data(instances)
-            dataset = instances[0].dataset.as_tiled_full(
-                image_data_list,
-                scale
-            )
-
-            with WsiDicomFileWriter(filepath) as wsi_file:
-                wsi_file.write(
-                    uid,
-                    transfer_syntax,
-                    dataset,
-                    image_data_list,
-                    workers,
-                    chunk_size,
-                    offset_table,
-                    scale
-                )
-            filepaths.append(filepath)
-
-        created_instances = WsiInstance.open(
-            [WsiDicomFile(filepath) for filepath in filepaths],
-            self.uids,
-            self.tile_size
-        )
-        return WsiDicomLevel(created_instances, self._base_pixel_spacing)
-
-
-class WsiDicomSeries(metaclass=ABCMeta):
-    """Represents a series of WsiDicomGroups with the same image flavor, e.g.
-    pyramidal levels, lables, or overviews.
-    """
-
-    def __init__(self, groups: Sequence[WsiDicomGroup]):
-        """Create a WsiDicomSeries from list of WsiDicomGroups.
-
-        Parameters
-        ----------
-        groups: Sequence[WsiDicomGroup]
-            List of groups to include in the series.
-        """
-        self._groups = groups
-
-        if len(self.groups) != 0 and self.groups[0].uids is not None:
-            self._uids = self._validate_series(self.groups)
-        else:
-            self._uids = None
-
-    def __repr__(self) -> str:
-        return f"{type(self).__name__}({self.groups})"
-
-    def __str__(self) -> str:
-        return f"{type(self).__name__} of groups {self.groups}"
-
-    def __getitem__(self, index: int) -> WsiDicomGroup:
-        """Get group by index.
-
-        Parameters
-        ----------
-        index: int
-            Index in series to get
-
-        Returns
-        ----------
-        WsiDicomGroup
-            The group at index in the series
-        """
-        return self.groups[index]
-
-    def __len__(self) -> int:
-        return len(self.groups)
-
-    @property
-    @abstractmethod
-    def wsi_type(self) -> str:
-        """Should return the wsi type of the series ('VOLUME', 'LABEL', or
-        'OVERVIEW'"""
-        raise NotImplementedError()
-
-    @property
-    def groups(self) -> Sequence[WsiDicomGroup]:
-        """Return contained groups."""
-        return self._groups
-
-    @property
-    def uids(self) -> Optional[SlideUids]:
-        """Return uids."""
-        return self._uids
-
-    @property
-    def mpps(self) -> List[SizeMm]:
-        """Return contained mpp (um/px)."""
-        return [group.mpp for group in self.groups]
-
-    @property
-    def files(self) -> List[Path]:
-        """Return contained files."""
-        series_files = [series.files for series in self.groups]
-        return [file for sublist in series_files for file in sublist]
-
-    @property
-    def datasets(self) -> List[WsiDataset]:
-        """Return contained datasets."""
-
-        series_datasets = [
-            series.datasets for series in self.groups
-        ]
-        return [
-            dataset for sublist in series_datasets for dataset in sublist
-        ]
-
-    @property
-    def instances(self) -> List[WsiInstance]:
-        """Return contained instances"""
-        series_instances = [
-            series.instances.values() for series in self.groups
-        ]
-        return [
-            instance for sublist in series_instances for instance in sublist
-        ]
-
-    @classmethod
-    @abstractmethod
-    def open(cls, instances: Sequence[WsiInstance]) -> 'WsiDicomSeries':
-        raise NotImplementedError
-
-    def _validate_series(
-            self,
-            groups: Union[Sequence[WsiDicomGroup], Sequence[WsiDicomLevel]]
-    ) -> Optional[SlideUids]:
-        """Check that no files or instances in series is duplicate and that
-        all groups in series matches.
-        Raises WsiDicomMatchError otherwise.
-        Returns the matching base uid. If list of groups is empty, return None.
-
-        Parameters
-        ----------
-        groups: Union[Sequence[WsiDicomGroup], Sequence[WsiDicomLevel]]
-            List of groups or levels to check
-
-        Returns
-        ----------
-        Optional[SlideUids]:
-            Matching uids
-        """
-        WsiDataset.check_duplicate_dataset(self.datasets, self)
-        WsiInstance.check_duplicate_instance(self.instances, self)
-
-        try:
-            base_group = groups[0]
-            if base_group.wsi_type != self.wsi_type:
-                raise WsiDicomMatchError(
-                    str(base_group), str(self)
-                )
-            for group in groups[1:]:
-                if not group.matches(base_group):
-                    raise WsiDicomMatchError(
-                        str(group), str(self)
-                    )
-            return base_group.uids
-        except IndexError:
-            return None
-
-    def close(self) -> None:
-        """Close all groups in the series."""
-        for group in self.groups:
-            group.close()
-
-    def save(
-        self,
-        output_path: str,
-        uid_generator: Callable[..., UID],
-        workers: int,
-        chunk_size: int,
-        offset_table: Optional[str]
-    ) -> List[Path]:
-        """Save WsiDicomSeries as DICOM-files in path.
-
-        Parameters
-        ----------
-        output_path: str
-        uid_generator: Callable[..., UID]
-             Function that can gernerate unique identifiers.
-        workers: int
-            Maximum number of thread workers to use.
-        chunk_size:
-            Chunk size (number of tiles) to process at a time. Actual chunk
-            size also depends on minimun_chunk_size from image_data.
-        offset_table: Optional[str] = 'bot'
-            Offset table to use, 'bot' basic offset table, 'eot' extended
-            offset table, None - no offset table.
-
-        Returns
-        ----------
-        List[Path]
-            List of paths of created files.
-        """
-        filepaths: List[Path] = []
-        for group in self.groups:
-            group_file_paths = group.save(
-                output_path,
-                uid_generator,
-                workers,
-                chunk_size,
-                offset_table
-            )
-            filepaths.extend(group_file_paths)
-        return filepaths
-
-
-class WsiDicomLabels(WsiDicomSeries):
-    """Represents a series of WsiDicomGroups of the label wsi flavor."""
-    WSI_TYPE = 'LABEL'
-
-    @property
-    def wsi_type(self) -> str:
-        return self.WSI_TYPE
-
-    @classmethod
-    def open(
-        cls,
-        instances: Sequence[WsiInstance]
-    ) -> 'WsiDicomLabels':
-        """Return labels created from wsi files.
-
-        Parameters
-        ----------
-        instances: Sequence[WsiInstance]
-            Instances to create labels from.
-
-        Returns
-        ----------
-        WsiDicomOverviews
-            Created labels.
-        """
-        labels = WsiDicomGroup.open(instances)
-        return cls(labels)
-
-
-class WsiDicomOverviews(WsiDicomSeries):
-    """Represents a series of WsiDicomGroups of the overview wsi flavor."""
-    WSI_TYPE = 'OVERVIEW'
-
-    @property
-    def wsi_type(self) -> str:
-        return self.WSI_TYPE
-
-    @classmethod
-    def open(
-        cls,
-        instances: Sequence[WsiInstance]
-    ) -> 'WsiDicomOverviews':
-        """Return overviews created from wsi files.
-
-        Parameters
-        ----------
-        instances: Sequence[WsiInstance]
-            Instances to create overviews from.
-
-        Returns
-        ----------
-        WsiDicomOverviews
-            Created overviews.
-        """
-        overviews = WsiDicomGroup.open(instances)
-        return cls(overviews)
-
-
-class WsiDicomLevels(WsiDicomSeries):
-    """Represents a series of WsiDicomGroups of the volume (e.g. pyramidal
-    level) wsi flavor."""
-    WSI_TYPE = 'VOLUME'
-
-    @property
-    def wsi_type(self) -> str:
-        return self.WSI_TYPE
-
-    @classmethod
-    def open(
-        cls,
-        instances: Sequence[WsiInstance]
-    ) -> 'WsiDicomLevels':
-        """Return overviews created from wsi files.
-
-        Parameters
-        ----------
-        instances: Sequence[WsiInstance]
-            Instances to create levels from.
-
-        Returns
-        ----------
-        WsiDicomOverviews
-            Created levels.
-        """
-        levels = WsiDicomLevel.open(instances)
-        return cls(levels)
-
-    def __init__(self, levels: Sequence[WsiDicomLevel]):
-        """Holds a stack of levels.
-
-        Parameters
-        ----------
-        levels: Sequence[WsiDicomLevel]
-            List of levels to include in series
-        """
-        self._levels = OrderedDict(
-            (level.level, level)
-            for level in sorted(levels, key=lambda level: level.level)
-        )
-        if len(self.groups) != 0 and self.groups[0].uids is not None:
-            self._uids = self._validate_series(self.groups)
-        else:
-            self._uids = None
-
-        mm_size = self.base_level.default_instance.mm_size
-        if mm_size is None:
-            raise ValueError(
-                'ImagedVolumeWidth and ImagedVolumeHeight must be set for '
-                '"Volume" type'
-            )
-        self._mm_size = mm_size
-
-    @property
-    def pyramid(self) -> str:
-        """Return string representation of pyramid"""
-        return (
-            'Pyramid levels in file:\n'
-            + '\n'.join(
-                [str_indent(2) + level.pyramid
-                 for level in self._levels.values()]
-            )
-        )
-
-    @property
-    def groups(self) -> List[WsiDicomGroup]:
-        """Return contained groups"""
-        return list(self._levels.values())
-
-    @property
-    def levels(self) -> List[int]:
-        """Return contained levels"""
-        return list(self._levels.keys())
-
-    @property
-    def highest_level(self) -> int:
-        """Return highest valid pyramid level (which results in a 1x1 image)"""
-        return self.base_level.get_highest_level()
-
-    @property
-    def base_level(self) -> WsiDicomLevel:
-        """Return the base level of the pyramid"""
-        return self._levels[0]
-
-    @property
-    def mm_size(self) -> SizeMm:
-        return self._mm_size
-
-    @property
-    def image_origin(self) -> ImageOrgin:
-        return self.base_level.default_instance.image_origin
-
-    def valid_level(self, level: int) -> bool:
-        """Check that given level is less or equal to the highest level
-        (1x1 pixel level).
-
-        Parameters
-        ----------
-        level: int
-            The level to check
-
-        Returns
-        ----------
-        bool
-            True if level is valid
-        """
-        return level <= self.highest_level
-
-    def get_level(self, level: int) -> WsiDicomLevel:
-        """Return wsi level.
-
-        Parameters
-        ----------
-        level: int
-            The level of the wsi level to return
-
-        Returns
-        ----------
-        WsiDicomLevel
-            The searched level
-        """
-        try:
-            return self._levels[level]
-        except KeyError:
-            raise WsiDicomNotFoundError(
-                f"Level of {level}", "level series"
-            )
-
-    def get_closest_by_level(self, level: int) -> WsiDicomLevel:
-        """Search for level that is closest to and smaller than the given
-        level.
-
-        Parameters
-        ----------
-        level: int
-            The level to search for
-
-        Returns
-        ----------
-        WsiDicomLevel
-            The level closest to searched level
-        """
-        if not self.valid_level(level):
-            raise WsiDicomOutOfBoundsError(
-                f"Level {level}", f"maximum level {self.highest_level}"
-            )
-        closest_level = 0
-        closest = None
-        for wsi_level in self._levels.values():
-            if((level >= wsi_level.level) and
-               (closest_level <= wsi_level.level)):
-                closest_level = wsi_level.level
-                closest = wsi_level
-        if closest is None:
-            raise WsiDicomNotFoundError(
-                f"Level for {level}", "level series"
-            )
-        return closest
-
-    def get_closest_by_size(self, size: Size) -> WsiDicomLevel:
-        """Search for level that by size is closest to and larger than the
-        given size.
-
-        Parameters
-        ----------
-        size: Size
-            The size to search for
-
-        Returns
-        ----------
-        WsiDicomLevel
-            The level with size closest to searched size
-        """
-        closest_size = self.groups[0].size
-        closest = None
-        for wsi_level in self._levels.values():
-            if((size.width <= wsi_level.size.width) and
-               (wsi_level.size.width <= closest_size.width)):
-                closest_size = wsi_level.size
-                closest = wsi_level
-        if closest is None:
-            raise WsiDicomNotFoundError(
-                f"Level for size {size}", "level series"
-            )
-        return closest
-
-    def get_closest_by_pixel_spacing(
-        self,
-        pixel_spacing: SizeMm
-    ) -> WsiDicomLevel:
-        """Search for level that by pixel spacing is closest to and smaller
-        than the given pixel spacing. Only the spacing in x-axis is used.
-
-        Parameters
-        ----------
-        pixel_spacing: SizeMm
-            Pixel spacing to search for
-
-        Returns
-        ----------
-        WsiDicomLevel
-            The level with pixel spacing closest to searched spacing
-        """
-        closest_pixel_spacing: float = 0
-        closest = None
-        for wsi_level in self._levels.values():
-            if((pixel_spacing.width >= wsi_level.pixel_spacing.width) and
-               (closest_pixel_spacing <= wsi_level.pixel_spacing.width)):
-                closest_pixel_spacing = wsi_level.pixel_spacing.width
-                closest = wsi_level
-        if closest is None:
-            raise WsiDicomNotFoundError(
-                f"Level for pixel spacing {pixel_spacing}", "level series")
-        return closest
-
-    def construct_pyramid(
-        self,
-        highest_level: int,
-        uid_generator: Callable[..., UID] = generate_uid,
-        workers: Optional[int] = None,
-        chunk_size: int = 100,
-        offset_table: Optional[str] = 'bot',
-        add_to_excisting: bool = True
-    ) -> List[Path]:
-        """Construct missing pyramid levels from excisting levels.
-
-        Parameters
-        ----------
-        highest_level: int
-        uid_generator: Callable[..., UID] = pydicom.uid.generate_uid
-             Function that can gernerate unique identifiers.
-        workers: Optional[int] = None
-            Maximum number of thread workers to use.
-        chunk_size: int = 100
-            Chunk size (number of tiles) to process at a time. Actual chunk
-            size also depends on minimun_chunk_size from image_data.
-        offset_table: Optional[str] = 'bot'
-            Offset table to use, 'bot' basic offset table, 'eot' extended
-            offset table, None - no offset table.
-        add_to_excisting: bool = True
-            If to add the created levels to excisting levels.
-
-        Returns
-        ----------
-        List[Path]
-            List of paths of created files.
-        """
-        if workers is None:
-            cpus = os.cpu_count()
-            if cpus is None:
-                workers = 1
-            else:
-                workers = cpus
-
-        filepaths: List[Path] = []
-
-        for pyramid_level in range(highest_level):
-            if pyramid_level not in self._levels.keys():
-                # Find the closest larger level for missing level
-                closest_level = self.get_closest_by_level(pyramid_level)
-                # Create scaled level
-                output_path = closest_level.files[0].parent
-                new_level = closest_level.create_child(
-                    scale=2,
-                    output_path=output_path,
-                    uid_generator=uid_generator,
-                    workers=workers,
-                    chunk_size=chunk_size,
-                    offset_table=offset_table
-                )
-                # Add level to available levels
-                if add_to_excisting:
-                    self._levels[new_level.level] = new_level
-                else:
-                    new_level.close()
-                filepaths += new_level.files
-        return filepaths
 
 
 class WsiDicom:
@@ -1453,20 +41,20 @@ class WsiDicom:
     labels and/or overviews."""
     def __init__(
         self,
-        levels: WsiDicomLevels,
-        labels: WsiDicomLabels,
-        overviews: WsiDicomOverviews,
+        levels: Levels,
+        labels: Labels,
+        overviews: Overviews,
         annotations: Optional[Sequence[AnnotationInstance]] = None
     ):
         """Holds wsi dicom levels, labels and overviews.
 
         Parameters
         ----------
-        levels: WsiDicomLevels
+        levels: Levels
             Series of pyramidal levels.
-        labels: WsiDicomLabels
+        labels: Labels
             Series of label images.
-        overviews: WsiDicomOverviews
+        overviews: Overviews
             Series of overview images
         annotations: Optional[Sequence[AnnotationInstance]] = None
             Sup-222 annotation instances.
@@ -1509,7 +97,7 @@ class WsiDicom:
         return self.pretty_str()
 
     @property
-    def base_level(self) -> WsiDicomLevel:
+    def base_level(self) -> Level:
         return self.levels.base_level
 
     @property
@@ -1523,21 +111,21 @@ class WsiDicom:
         return self.base_level.tile_size
 
     @property
-    def levels(self) -> WsiDicomLevels:
+    def levels(self) -> Levels:
         """Return contained levels"""
         if self._levels is not None:
             return self._levels
         raise WsiDicomNotFoundError("levels", str(self))
 
     @property
-    def labels(self) -> WsiDicomLabels:
+    def labels(self) -> Labels:
         """Return contained labels"""
         if self._labels is not None:
             return self._labels
         raise WsiDicomNotFoundError("labels", str(self))
 
     @property
-    def overviews(self) -> WsiDicomOverviews:
+    def overviews(self) -> Overviews:
         """Return contained overviews"""
         if self._overviews is not None:
             return self._overviews
@@ -1549,7 +137,7 @@ class WsiDicom:
         return self.levels.files + self.labels.files + self.overviews.files
 
     @property
-    def datasets(self) -> List[WsiDataset]:
+    def datasets(self) -> List[WsiDicomDataset]:
         """Return contained datasets."""
         return (
             self.levels.datasets
@@ -1558,7 +146,7 @@ class WsiDicom:
         )
 
     @property
-    def instances(self) -> List[WsiInstance]:
+    def instances(self) -> List[Instance]:
         """Return contained instances"""
         return (
             self.levels.instances
@@ -1622,11 +210,11 @@ class WsiDicom:
             sop_class_uid = cls._get_sop_class_uid(filepath)
             if sop_class_uid == WSI_SOP_CLASS_UID:
                 wsi_file = WsiDicomFile(filepath)
-                if(wsi_file.wsi_type == WsiDicomLevels.WSI_TYPE):
+                if(wsi_file.wsi_type == Levels.WSI_TYPE):
                     level_files.append(wsi_file)
-                elif(wsi_file.wsi_type == WsiDicomLabels.WSI_TYPE):
+                elif(wsi_file.wsi_type == Labels.WSI_TYPE):
                     label_files.append(wsi_file)
-                elif(wsi_file.wsi_type == WsiDicomOverviews.WSI_TYPE):
+                elif(wsi_file.wsi_type == Overviews.WSI_TYPE):
                     overview_files.append(wsi_file)
                 else:
                     wsi_file.close()
@@ -1635,17 +223,17 @@ class WsiDicom:
         base_dataset = cls._get_base_dataset(level_files)
         slide_uids = base_dataset.uids.slide
         base_tile_size = base_dataset.tile_size
-        level_instances = WsiInstance.open(
+        level_instances = Instance.open(
             level_files,
             slide_uids,
             base_tile_size
         )
-        label_instances = WsiInstance.open(label_files, slide_uids)
-        overview_instances = WsiInstance.open(overview_files, slide_uids)
+        label_instances = Instance.open(label_files, slide_uids)
+        overview_instances = Instance.open(overview_files, slide_uids)
 
-        levels = WsiDicomLevels.open(level_instances)
-        labels = WsiDicomLabels.open(label_instances)
-        overviews = WsiDicomOverviews.open(overview_instances)
+        levels = Levels.open(level_instances)
+        labels = Labels.open(label_instances)
+        overviews = Overviews.open(overview_instances)
         annotations = AnnotationInstance.open(annotation_files)
 
         return cls(levels, labels, overviews, annotations)
@@ -1941,7 +529,7 @@ class WsiDicom:
         level: int,
         z: Optional[float] = None,
         path: Optional[str] = None
-    ) -> WsiInstance:
+    ) -> Instance:
 
         """Return instance fullfilling level, z and/or path.
 
@@ -1956,7 +544,7 @@ class WsiDicom:
 
         Returns
         ----------
-        WsiInstance:
+        Instance:
             Instance
         """
         wsi_level = self.levels.get_level(level)
@@ -2009,7 +597,7 @@ class WsiDicom:
         if chunk_size is None:
             chunk_size = 16
 
-        collections: List[WsiDicomSeries] = [
+        collections: List[Series] = [
             self.levels, self.labels, self.overviews
         ]
 
@@ -2067,7 +655,7 @@ class WsiDicom:
     @staticmethod
     def _get_base_dataset(
         files: Sequence[WsiDicomFile]
-    ) -> WsiDataset:
+    ) -> WsiDicomDataset:
         """Return file with largest image (width) from list of files.
 
         Parameters
@@ -2077,7 +665,7 @@ class WsiDicom:
 
         Returns
         ----------
-        WsiDataset
+        WsiDicomDataset
             Base layer dataset.
         """
         base_size = Size(0, 0)
@@ -2108,7 +696,7 @@ class WsiDicom:
 
     def _validate_collection(
         self,
-        series: Sequence[WsiDicomSeries]
+        series: Sequence[Series]
     ) -> SlideUids:
         """Check that no files or instance in collection is duplicate, and, if
         strict, that all series have the same base uids.
@@ -2116,7 +704,7 @@ class WsiDicom:
 
         Parameters
         ----------
-        series: Sequence[WsiDicomSeries]
+        series: Sequence[Series]
             List of series to check.
 
         Returns
@@ -2124,8 +712,8 @@ class WsiDicom:
         SlideUids
             Matching uids
         """
-        WsiDataset.check_duplicate_dataset(self.datasets, self)
-        WsiInstance.check_duplicate_instance(self.instances, self)
+        WsiDicomDataset.check_duplicate_dataset(self.datasets, self)
+        Instance.check_duplicate_instance(self.instances, self)
 
         try:
             slide_uids = next(
